requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiofiles==25.1.0
    # via backend (pyproject.toml)
alembic==1.18.1
    # via backend (pyproject.toml)
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
    #   watchfiles
argon2-cffi==25.1.0
    # via backend (pyproject.toml)
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
asyncpg==0.31.0
    # via backend (pyproject.toml)
bcrypt==5.0.0
    # via backend (pyproject.toml)
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   sentry-sdk
cffi==2.0.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
click==8.3.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
cryptography==46.0.3
    # via backend (pyproject.toml)
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi==0.128.0
    # via backend (pyproject.toml)
fastapi-cli==0.0.20
    # via fastapi
fastapi-cloud-cli==0.11.0
    # via fastapi-cli
fastar==0.8.0
    # via fastapi-cloud-cli
greenlet==3.3.0
    # via backend (pyproject.toml)
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
iniconfig==2.3.0
    # via pytest
jinja2==3.1.6
    # via fastapi
mako==1.3.10
    # via alembic
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
packaging==25.0
    # via pytest
passlib==1.7.4
    # via backend (pyproject.toml)
pluggy==1.6.0
    # via pytest
psycopg==3.3.2
    # via backend (pyproject.toml)
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via
    #   backend (pyproject.toml)
    #   fastapi
    #   fastapi-cloud-cli
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-extra-types==2.11.0
    # via fastapi
pydantic-settings==2.12.0
    # via fastapi
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyjwt==2.10.1
    # via backend (pyproject.toml)
pytest==9.0.2
    # via backend (pyproject.toml)
python-dotenv==1.2.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.21
    # via fastapi
pyyaml==6.0.3
    # via uvicorn
rich==14.2.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.17.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.6
    # via fastapi-cloud-cli
sentry-sdk==2.50.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
sqlalchemy==2.0.45
    # via
    #   backend (pyproject.toml)
    #   alembic
starlette==0.50.0
    # via fastapi
typer==0.21.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
types-aiofiles==25.1.0.20251011
    # via backend (pyproject.toml)
types-cryptography==3.3.23.2
    # via types-pyjwt
types-passlib==1.7.7.20250602
    # via backend (pyproject.toml)
types-pyjwt==1.7.1
    # via backend (pyproject.toml)
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich-toolkit
    #   sqlalchemy
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via sentry-sdk
uvicorn==0.40.0
    # via
    #   backend (pyproject.toml)
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
